use static JSON for characters
