+++ .editorconfig
@@ -7,6 +7,3 @@ indent_size = 2
 end_of_line = lf
 insert_final_newline = true
 trim_trailing_whitespace = true
-
-[*.css]
-indent_style = tab
