+++ package.json
@@ -21,8 +21,6 @@
   },
   "dependencies": {
     "botkit": "0.2.2",
-    "cheerio": "0.22.0",
-    "cheerio-tableparser": "1.0.1",
     "winston": "2.2.0"
   },
   "engines": {
